+++ src/plugins/valueHandler/basic/basicValueHandler.component.ts
@@ -9,6 +9,8 @@ import {ɵNgSelectOption, } from '../../../components/option';
 import {ValueHandlerBase} from '../valueHandlerBase';
 import {PluginBus} from '../../../misc/pluginBus/pluginBus';
 
+//TODO - select _setValue, rejest non existing
+
 /**
  * Default options for value handler
  * @internal